+++ PLBase/PLGeneral/include/PLGeneral/System/SemaphoreLinux.h
@@ -80,7 +80,7 @@ class SemaphoreLinux : public SemaphoreImpl {
 	//[-------------------------------------------------------]
 	private:
 		virtual bool Lock();
-		virtual bool TryLock(uint32 nTimeout);
+		virtual bool TryLock(uint64 nTimeout);
 		virtual bool Unlock();
