+++ PLBase/PLGeneral/include/PLGeneral/System/SemaphoreWindows.h
@@ -81,7 +81,7 @@ class SemaphoreWindows : public SemaphoreImpl {
 	//[-------------------------------------------------------]
 	private:
 		virtual bool Lock();
-		virtual bool TryLock(uint32 nTimeout);
+		virtual bool TryLock(uint64 nTimeout);
 		virtual bool Unlock();
