+++ PLBase/PLGeneral/src/System/SemaphoreWindows.cpp
@@ -65,10 +65,10 @@ bool SemaphoreWindows::Lock()
 	return (m_hSemaphore && WaitForSingleObject(m_hSemaphore, INFINITE) == WAIT_OBJECT_0);
 }
 
-bool SemaphoreWindows::TryLock(uint32 nTimeout)
+bool SemaphoreWindows::TryLock(uint64 nTimeout)
 {
 	// Lock semaphore
-	return (m_hSemaphore && WaitForSingleObject(m_hSemaphore, nTimeout) == WAIT_OBJECT_0);
+	return (m_hSemaphore && WaitForSingleObject(m_hSemaphore, static_cast<DWORD>(nTimeout)) == WAIT_OBJECT_0);
 }
 
 bool SemaphoreWindows::Unlock()
